source_gen: fix another Apple-Clang warning.

+++ benchmarks/multi_normal.cpp
@@ -203,8 +203,8 @@ namespace {
         at::Tensor minimum_L = torch::full(
             {matrix_size, matrix_size}, -inf, options
         );
-        for(size_t j = 0; j < matrix_size; ++j) {
-            minimum_L[j][int64_t(j)] = 1e-1;
+        for(int64_t j = 0; j < int64_t(matrix_size); ++j) {
+            minimum_L[j][j] = 1e-1;
         }
         minimum_L = torch::tril( minimum_L );
         return minimum_L;
